pyproject.toml
[project]
name = "health-visualizer"
version = "0.1.0"
description = ""
authors = [{ name = "Mike Guida", email = "mike@mguida.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["foxglove-sdk (>=0.4.0,<0.5.0)", "gpxpy (>=1.6.2,<2.0.0)", "geojson (>=3.2.0,<4.0.0)", "pytz (>=2025.2,<2026.0)"]

[project.scripts]
health_visualizer = "health_visualizer.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

